moved last two tests
