+++ apps/backend-serverless/__tests__/unit/models/dependencies/trm-wallet-screen.model.test.ts
@@ -1,4 +1,4 @@
-import { parseAndValidateTrmWalletScreenResponse } from '../../../src/models/dependencies/trm-wallet-screen-response.model.js';
+import { parseAndValidateTrmWalletScreenResponse } from '../../../../src/models/dependencies/trm-wallet-screen-response.model.js';
 
 describe('unit testing the trm wallet screen model', () => {
     it('valid trm wallet screen response body', () => {
